Refactor + cleanup. Add api-json-docs folder for reference.
