+++ app/api/hotels/search/route.ts
@@ -1,11 +1,5 @@
-import {
-  validateDateFormat,
-  validateDateRange,
-} from "@/lib/rapid-hotel-api/validateDates";
-import {
-  validateDomain,
-  validateLocale,
-} from "@/lib/rapid-hotel-api/validateDomainLocale";
+import { validateDateFormatAndDateRange } from "@/lib/rapid-hotel-api/validateDates";
+import { validateDomainAndLocale } from "@/lib/rapid-hotel-api/validateDomainLocale";
 import { validatePriceRange } from "@/lib/rapid-hotel-api/validatePriceRange";
 import { API_OPTIONS } from "@/types/rapid-hotels-api/api-types";
 import {
@@ -49,7 +43,7 @@ function validateSearchParams(searchParams: URLSearchParams) {
     "meal_plan",
     "available_filter",
   ];
-  const errors: string[] = [];
+  let errors: string[] = [];
 
   // Validate required searchParams and check for missing searchParams.
   requiredSearchParams.forEach((searchParam) => {
@@ -61,19 +55,8 @@ function validateSearchParams(searchParams: URLSearchParams) {
   // Validate check-in and check-out date formatting/range.
   const checkinDate = searchParams.get("checkin_date");
   const checkoutDate = searchParams.get("checkout_date");
-  const checkinDateFormatError = validateDateFormat(
-    checkinDate,
-    "checkin_date"
-  );
-  const checkoutDateFormatError = validateDateFormat(
-    checkoutDate,
-    "checkout_date"
-  );
-  if (checkinDateFormatError) errors.push(checkinDateFormatError);
-  if (checkoutDateFormatError) errors.push(checkoutDateFormatError);
-  // Validate date range between check-in and check-out dates
-  const dateRangeError = validateDateRange(checkinDate, checkoutDate);
-  if (dateRangeError) errors.push(dateRangeError);
+  const dateErrors = validateDateFormatAndDateRange(checkinDate, checkoutDate);
+  if (dateErrors) errors = [...errors, ...dateErrors];
 
   // Validate min and max price range.
   const minPrice = searchParams.get("price_min");
@@ -82,10 +65,11 @@ function validateSearchParams(searchParams: URLSearchParams) {
   if (priceRangeError) errors.push(priceRangeError);
 
   // Validate domain and locale
-  const domainError = validateDomain(requiredWithDefaultSearchParams.domain);
-  const localeError = validateLocale(requiredWithDefaultSearchParams.locale);
-  if (domainError) errors.push(domainError);
-  if (localeError) errors.push(localeError);
+  const domainLocaleErrors = validateDomainAndLocale(
+    requiredWithDefaultSearchParams.domain,
+    requiredWithDefaultSearchParams.locale
+  );
+  if (domainLocaleErrors) errors = [...errors, ...domainLocaleErrors];
 
   // Check for errors in required searchParams
   if (errors.length > 0) {
@@ -174,222 +158,3 @@ export async function GET(req: NextRequest) {
     );
   }
 }
-
-// Refer to types/rapid-hotels-api/hotel-search-types.ts for API input searchParams. See HotelsSearchQuery.\
-// Otherwise, check possible searchParams in requiredSearchParams, requiredWithDefaultSearchParams (required, but gives default value if not provided), optionalSearchParams in function validateSearchParams.
-
-// Hotel Search API output:
-
-// Level 0 - API output
-type PropertySearchResults = {
-  __typename: "PropertySearchResults";
-  filterMetadata: PropertyFilterMetadata;
-  properties: Property[];
-  universalSortAndFilter: {}; // Placeholder, as details are not required
-  propertySearchListings: {}; // Placeholder, as details are not required
-  summary: {}; // Placeholder, as details are not required
-  searchCriteria: {}; // Placeholder, as details are not required
-  shoppingContext: {}; // Placeholder, as details are not required
-  map: {}; // Placeholder, as details are not required
-  clickstream: {}; // Placeholder, as details are not required
-  [key: string]: any;
-};
-
-// Level 1 - filterMetadata
-type PropertyFilterMetadata = {
-  __typename: "PropertyFilterMetadata";
-  priceRange: PriceRange; // Set minPrice and maxPrice absolute limits.
-  [key: string]: any;
-};
-// Level 2 - filterMetadata attributes
-type PriceRange = {
-  __typename: "PriceRange";
-  max: number;
-  min: number;
-  [key: string]: any;
-};
-
-// Level 1 - properties (hotels)
-type Property = {
-  __typename: "Property";
-  id: string; // hotelId, required for HotelRoomOffers api
-  name: string;
-  featuredMessages: any[];
-  availability: PropertyAvailability;
-  propertyImage: PropertyImage;
-  destinationInfo: PropertyDestinationInfo;
-  legalDisclaimer: string | null;
-  listingFooter: string | null;
-  mapMarker: MapMarker;
-  neighborhood: Region;
-  offerBadge: string | null;
-  offerSummary: OfferSummary;
-  pinnedDetails: string | null;
-  price: PropertyPrice; // Price of hotel
-  priceAfterLoyaltyPointsApplied: PropertyPrice;
-  propertyFees: any[];
-  reviews: PropertyReviewsSummary;
-  star: number | null;
-  supportingMessages: string | null;
-  regionId: string;
-  priceMetadata: PropertyPriceMetadata;
-  saveTripItem: string | null;
-  [key: string]: any;
-};
-
-// Level 2 - Property (hotel) attributes
-type PropertyAvailability = {
-  __typename: "PropertyAvailability";
-  available: boolean;
-  minRoomsLeft: number | null;
-};
-
-type PropertyImage = {
-  __typename: "PropertyImage";
-  alt: string;
-  fallbackImage: string | null;
-  image: Image;
-  subjectId: number;
-};
-
-type Image = {
-  __typename: "Image";
-  description: string;
-  url: string;
-};
-
-type PropertyDestinationInfo = {
-  __typename: "PropertyDestinationInfo";
-  distanceFromDestination: Distance;
-  distanceFromMessaging: string;
-  regionId: string;
-};
-
-type Distance = {
-  __typename: "Distance";
-  unit: string;
-  value: number;
-};
-
-type MapMarker = {
-  __typename: "MapMarker";
-  label: string;
-  latLong: Coordinates;
-};
-
-type Coordinates = {
-  __typename: "Coordinates";
-  latitude: number;
-  longitude: number;
-};
-
-type Region = {
-  __typename: "Region";
-  name: string;
-};
-
-type OfferSummary = {
-  __typename: "OfferSummary";
-  messages: OfferSummaryMessage[];
-  attributes: OfferAttribute[];
-};
-
-type OfferSummaryMessage = {
-  __typename: "OfferSummaryMessage";
-  message: string;
-  theme: string;
-  type: string;
-  mark: string | null;
-};
-
-type OfferAttribute = {
-  __typename: "OfferAttribute";
-  type: string;
-};
-
-type PropertyPrice = {
-  __typename: "PropertyPrice";
-  options: PropertyPriceOption[];
-  priceMessaging: string | null;
-  lead: Money;
-  strikeOut: Money;
-  displayMessages: PriceDisplayMessage[];
-  strikeOutType: string;
-  priceMessages: LodgingPlainMessage[];
-};
-
-type PropertyPriceOption = {
-  __typename: "PropertyPriceOption";
-  strikeOut: Money;
-  disclaimer: LodgingPlainMessage;
-  formattedDisplayPrice: string;
-};
-
-type Money = {
-  __typename: "Money";
-  amount: number;
-  formatted: string;
-  currencyInfo?: Currency; // Optional if needed in some contexts
-};
-
-type Currency = {
-  __typename: "Currency";
-  code: string;
-  symbol: string;
-};
-
-type PriceDisplayMessage = {
-  __typename: "PriceDisplayMessage";
-  lineItems: (LodgingEnrichedMessage | DisplayPrice)[];
-};
-
-type LodgingEnrichedMessage = {
-  __typename: "LodgingEnrichedMessage";
-  accessibilityLabel: string | null;
-  mark: string | null;
-  state: string;
-  value: string;
-  badge: string | null;
-};
-
-type DisplayPrice = {
-  __typename: "DisplayPrice";
-  disclaimer: LodgingPlainDialog | null;
-  price: FormattedMoney;
-  role: string;
-};
-
-type LodgingPlainDialog = {
-  __typename: "LodgingPlainDialog";
-  content: string[];
-  title: string | null;
-};
-
-type FormattedMoney = {
-  __typename: "FormattedMoney";
-  formatted: string;
-  accessibilityLabel: string;
-};
-
-type LodgingPlainMessage = {
-  __typename: "LodgingPlainMessage";
-  value: string;
-};
-
-type PropertyReviewsSummary = {
-  __typename: "PropertyReviewsSummary";
-  score: number;
-  total: number;
-};
-
-type PropertyPriceMetadata = {
-  __typename: "PropertyPriceMetadata";
-  discountType: string;
-  rateDiscount: RateDiscount;
-  totalDiscountPercentage: number;
-};
-
-type RateDiscount = {
-  __typename: "RateDiscount";
-  description: string;
-};
