+++ types/rapid-hotels-api/api-json-docs/hotel-room-offers-doc.ts
@@ -1,0 +1,6 @@
+// Hotel Room Offers API
+// https://rapidapi.com/tipsters/api/hotels-com-provider/playground/apiendpoint_513f7e8b-520e-465d-8aca-f50cc42e6969
+
+// api/hotels/search/rooms API route
+// Find input in hotel-room-offers-types.ts
+// Output use postman API client or use rapidAPI link above to see structure
